requirements.txt
opencv-python~=4.6.0.66
mediapipe~=0.9.0
pyserial~=3.5